requirements.txt
pyserial>=3.5
hid>=1.0.4
micropython-cpython-ustruct==0.0
micropython-cpython-micropython==0.1.1
micropython-cpython-utime==0.4
